Update episodes and player to be components
