+++ app/templates/application.hbs
@@ -23,6 +23,6 @@
 
 <audio id="audio-player" preload="auto" mozaudiochannel="content"></audio>
 
-{{render "player"}}
+{{audio-player}}
 
 {{outlet "modal"}}
