+++ app/templates/components/episode-list.hbs
@@ -1,5 +1,5 @@
 <div class="episodes ui divided relaxed huge list">
-  {{#each model as |episode|}}
+  {{#each sortedEpisodes as |episode|}}
     <div class="item episode {{if isNew "new" "old"}}">
       <div class="content">
         <!--
